pyproject.toml
[tool.poetry]
name = "django-drf"
version = "1.0"
description = "Проект для производственной практики по DRF"
authors = ["Фирсанов <firsanovstanislav@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
